+++ docs/CHANGELOG.md
@@ -279,3 +279,7 @@ The switch to python allows cross platform support, standardized processes, and 
 ### 3.2.4
 
 1. Fix restart-server command in CLI
+
+### 3.2.5
+
+1. Fix download server function to use updated minecraft site